# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
      "AspectRatio": {
        "enum": [
          "1:1",
          "4:3",
          "3:4",
          "16:9"
        ],
        "type": "string"
      },
      "Body_create_image_edit_job_api_v1_image_edit_jobs_post": {
        "properties": {
          "aspectRatio": {
            "$ref": "#/components/schemas/AspectRatio"
          },
          "faceReference": {
            "anyOf": [
              {
                "contentMediaType": "application/octet-stream",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Facereference"
          },
          "prompt": {
            "maxLength": 1200,
            "minLength": 1,
            "title": "Prompt",
            "type": "string"
          },
          "seed": {
            "anyOf": [
              {
                "maximum": 2147483647.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed"
          },
          "source": {
            "contentMediaType": "application/octet-stream",
            "title": "Source",
            "type": "string"
          },
          "style": {
            "$ref": "#/components/schemas/ImageStyle"
          }
        },
        "required": [
          "prompt",
          "aspectRatio",
          "style",
          "source"
        ],
        "title": "Body_create_image_edit_job_api_v1_image_edit_jobs_post",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HealthResponse": {
        "properties": {
          "demoMode": {
            "title": "Demomode",
            "type": "boolean"
          },
          "service": {
            "const": "canvasrelay-api",
            "title": "Service",
            "type": "string"
          },
          "status": {
            "const": "ok",
            "title": "Status",
            "type": "string"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "status",
          "service",
          "version",
          "demoMode",
          "timestamp"
        ],
        "title": "HealthResponse",
        "type": "object"
      },
      "ImageJobCreate": {
        "properties": {
          "aspectRatio": {
            "$ref": "#/components/schemas/AspectRatio"
          },
          "prompt": {
            "maxLength": 1200,
            "minLength": 1,
            "title": "Prompt",
            "type": "string"
          },
          "seed": {
            "anyOf": [
              {
                "maximum": 2147483647.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed"
          },
          "style": {
            "$ref": "#/components/schemas/ImageStyle"
          }
        },
        "required": [
          "prompt",
          "aspectRatio",
          "style"
        ],
        "title": "ImageJobCreate",
        "type": "object"
      },
      "ImageJobError": {
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message",
          "action",
          "retryable"
        ],
        "title": "ImageJobError",
        "type": "object"
      },
      "ImageJobListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ImageJobResponse"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "ImageJobListResponse",
        "type": "object"
      },
      "ImageJobOperation": {
        "enum": [
          "generate",
          "edit"
        ],
        "type": "string"
      },
      "ImageJobResponse": {
        "properties": {
          "completedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completedat"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ImageJobError"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "progress": {
            "anyOf": [
              {
                "maximum": 100.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress"
          },
          "prompt": {
            "title": "Prompt",
            "type": "string"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ImageJobResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "settings": {
            "$ref": "#/components/schemas/ImageJobSettings"
          },
          "startedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Startedat"
          },
          "status": {
            "$ref": "#/components/schemas/ImageJobStatus"
          }
        },
        "required": [
          "id",
          "status",
          "progress",
          "prompt",
          "settings",
          "createdAt",
          "startedAt",
          "completedAt",
          "result",
          "error"
        ],
        "title": "ImageJobResponse",
        "type": "object"
      },
      "ImageJobResult": {
        "properties": {
          "height": {
            "title": "Height",
            "type": "integer"
          },
          "mimeType": {
            "$ref": "#/components/schemas/ImageMimeType"
          },
          "url": {
            "title": "Url",
            "type": "string"
          },
          "width": {
            "title": "Width",
            "type": "integer"
          }
        },
        "required": [
          "url",
          "mimeType",
          "width",
          "height"
        ],
        "title": "ImageJobResult",
        "type": "object"
      },
      "ImageJobSettings": {
        "properties": {
          "aspectRatio": {
            "$ref": "#/components/schemas/AspectRatio"
          },
          "hasFaceReference": {
            "title": "Hasfacereference",
            "type": "boolean"
          },
          "operation": {
            "$ref": "#/components/schemas/ImageJobOperation"
          },
          "provider": {
            "$ref": "#/components/schemas/ImageProviderName"
          },
          "seed": {
            "title": "Seed",
            "type": "integer"
          },
          "style": {
            "$ref": "#/components/schemas/ImageStyle"
          }
        },
        "required": [
          "aspectRatio",
          "style",
          "seed",
          "provider",
          "operation",
          "hasFaceReference"
        ],
        "title": "ImageJobSettings",
        "type": "object"
      },
      "ImageJobStatus": {
        "enum": [
          "queued",
          "running",
          "completed",
          "failed",
          "canceled"
        ],
        "type": "string"
      },
      "ImageMimeType": {
        "enum": [
          "image/svg+xml",
          "image/png",
          "image/jpeg",
          "image/webp"
        ],
        "type": "string"
      },
      "ImageProviderName": {
        "enum": [
          "demo",
          "comfyui"
        ],
        "type": "string"
      },
      "ImageProviderResponse": {
        "properties": {
          "label": {
            "title": "Label",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "mode": {
            "enum": [
              "demo",
              "live"
            ],
            "title": "Mode",
            "type": "string"
          },
          "provider": {
            "$ref": "#/components/schemas/ImageProviderName"
          },
          "ready": {
            "title": "Ready",
            "type": "boolean"
          }
        },
        "required": [
          "provider",
          "mode",
          "label",
          "ready",
          "message"
        ],
        "title": "ImageProviderResponse",
        "type": "object"
      },
      "ImageStyle": {
        "enum": [
          "editorial",
          "product",
          "concept"
        ],
        "type": "string"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Public orchestration API for the CanvasRelay studio.",
    "title": "CanvasRelay API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/health": {
      "get": {
        "operationId": "health_api_v1_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health",
        "tags": [
          "system"
        ]
      }
    },
    "/api/v1/image-edit-jobs": {
      "post": {
        "operationId": "create_image_edit_job_api_v1_image_edit_jobs_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_image_edit_job_api_v1_image_edit_jobs_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Image Edit Job",
        "tags": [
          "image edit jobs"
        ]
      }
    },
    "/api/v1/image-jobs": {
      "get": {
        "operationId": "list_image_jobs_api_v1_image_jobs_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 24,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ImageJobStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "operation",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ImageJobOperation"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Operation"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageJobListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Image Jobs",
        "tags": [
          "image jobs"
        ]
      },
      "post": {
        "operationId": "create_image_job_api_v1_image_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImageJobCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Image Job",
        "tags": [
          "image jobs"
        ]
      }
    },
    "/api/v1/image-jobs/{job_id}": {
      "delete": {
        "operationId": "cancel_image_job_api_v1_image_jobs__job_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Image Job",
        "tags": [
          "image jobs"
        ]
      },
      "get": {
        "operationId": "get_image_job_api_v1_image_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Image Job",
        "tags": [
          "image jobs"
        ]
      }
    },
    "/api/v1/image-jobs/{job_id}/events": {
      "get": {
        "operationId": "stream_image_job_api_v1_image_jobs__job_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream Image Job",
        "tags": [
          "image jobs"
        ]
      }
    },
    "/api/v1/image-jobs/{job_id}/result": {
      "get": {
        "operationId": "get_image_job_result_api_v1_image_jobs__job_id__result_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Image Job Result",
        "tags": [
          "image jobs"
        ]
      }
    },
    "/api/v1/providers/image": {
      "get": {
        "operationId": "get_image_provider_api_v1_providers_image_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageProviderResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Image Provider",
        "tags": [
          "providers"
        ]
      }
    },
    "/api/v1/providers/image-edit": {
      "get": {
        "operationId": "get_image_edit_provider_api_v1_providers_image_edit_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageProviderResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Image Edit Provider",
        "tags": [
          "providers"
        ]
      }
    }
  }
}